Updated Sponsors
Updated the sponsor carousel to include the ones only on our instagram so far. I am not sure which sponsors are guaranteed but I put those for now. Also im not sure whether to put the ones listed as vendor + sponsor because the "logo" would just be an instagram image. If there is a list of sponsors needed on the sponsor page that would be great.